Update file paths + Extract some test doubles
