+++ Tests/Support/Doubles/BodyThrowErrorFakeBuilder.swift
@@ -1,0 +1,10 @@
+import Foundation
+@testable import Frisbee
+
+struct BodyThrowErrorFakeBuilder: BodyBuildable {
+    private let errorToThrow = FrisbeeError.invalidEntity
+
+    func build<Entity: Encodable>(withBody body: Entity) throws -> [String: Any] {
+        throw errorToThrow
+    }
+}
