+++ Tests/Support/Entities/Fake.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct Fake: Codable {
+    let fake: String
+}
