+++ Tests/Support/Entities/SomeError.swift
